Refactor kubemodel: extract structs into separate files and move resources to KubeModelSet
- Extract all struct definitions from cluster.go into individual files:
   - provider.go, resolution.go, controller.go (with enums)
   - window.go, diagnostic.go
   - gpu_usage.go, gpu_device.go
   - container.go, namespace.go, service.go, node.go, pod.go
   - volume.go, pvc.go (using abbreviated naming)

- Move resource collections from Cluster struct to KubeModelSet as maps
  - Cluster now only contains cluster-level metadata
  - All resources (Containers, Controllers, GpuDevices, Namespaces, etc.)
    are now stored as maps in KubeModelSet for efficient lookups

- Implement namespace name-to-ID indexing
  - Add GetNamespaceByName() method for O(1) lookups by name
  - Update RegisterNamespace() to maintain the index

- Improve KubeModelSet initialization and methods
  - Fix NewKubeModelSet to use Resolution instead of end time
  - Implement IsEmpty() to check for nil, missing cluster, or no resources
  - Fix RegisterResourceQuota parameter handling
