+++ core/pkg/model/kubemodel/cluster.go
@@ -1,0 +1,9 @@
+package kubemodel
+
+// Cluster represents the top-level Kubernetes cluster
+type Cluster struct {
+	ID       string   `json:"id"`
+	Provider Provider `json:"provider"`
+	Account  string   `json:"account"`
+	Name     string   `json:"name"`
+}
