+++ core/pkg/model/kubemodel/window.go
@@ -1,0 +1,9 @@
+package kubemodel
+
+import "time"
+
+// Window defines a unit of time by a resolution and a start time
+type Window struct {
+	Resolution Resolution `json:"resolution"`
+	Start      time.Time  `json:"start"`
+}
